+++ src/storage/schema.rs
@@ -542,7 +542,14 @@ pub fn ddl_create_tt_proc_on_master(proc_name: &str) -> traft::Result<()> {
                 "cannot find procedure {proc_name} in `proc::all_procs` for schema creation"
             ))
         })?;
-    if sbroad::frontend::sql::NAMES_OF_FUNCTIONS_IN_SOURCES.contains(&proc_name) {
+
+    // This branching is needed because we have basically two "types" of exportable functions
+    // either we want maximum exposure (available from SQL through the wrapper), or we need
+    // them just internally, though exposed as C FFI functions only.
+    if sbroad::frontend::sql::FUNCTION_NAME_MAPPINGS
+        .iter()
+        .any(|mapping| mapping.rust_procedure == proc_name)
+    {
         lua.exec_with(
             "local name, is_public = ...
             local proc_name = '.' .. name